pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "easy-rfdetr"
version = "0.1.0"
description = "YOLO-like API for RF-DETR real-time object detection"
readme = "README.md"
authors = [
    {name = "easy-rfdetr", email = "dev@example.com"}
]
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "rfdetr>=1.0.0",
    "opencv-python>=4.8.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "rich>=13.0.0",
    "tqdm>=4.65.0",
    "requests>=2.28.0",
    "typer>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
gradio = [
    "gradio>=4.0.0",
]
youtube = [
    "yt-dlp>=2023.0.0",
]

[project.urls]
Homepage = "https://github.com/easy-rfdetr/easy-rfdetr"
Documentation = "https://easy-rfdetr.readthedocs.io"
Repository = "https://github.com/easy-rfdetr/easy-rfdetr"
Issues = "https://github.com/easy-rfdetr/easy-rfdetr/issues"

[project.scripts]
rfdetr = "easy_rfdetr.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["easy_rfdetr*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"
